Extract metadata tests
